HostプラグインのCanvasでOutOfMemoryが発生した時の処理を整理

+++ dConnectDevicePlugin/dConnectDeviceHost/app/src/main/java/org/deviceconnect/android/deviceplugin/host/profile/HostCanvasProfile.java
@@ -12,6 +12,7 @@ import android.app.Service;
 import android.content.Intent;
 import android.support.annotation.NonNull;
 import android.support.v4.content.LocalBroadcastManager;
+import android.util.Log;
 
 import org.deviceconnect.android.deviceplugin.host.activity.CanvasProfileActivity;
 import org.deviceconnect.android.deviceplugin.host.canvas.CanvasDrawImageObject;
@@ -25,9 +26,6 @@ import org.deviceconnect.message.DConnectMessage;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.Locale;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -39,13 +37,16 @@ import java.util.concurrent.Executors;
 public class HostCanvasProfile extends CanvasProfile {
     /** ファイルが生存できる有効時間を定義する. */
     private static final long DEFAULT_EXPIRE = 1000 * 60 * 5;
+
+    /** Canvasプロファイルのファイル名プレフィックス。 */
+    private static final String CANVAS_PREFIX = "host_canvas";
+
     /** ファイルが生存できる有効時間. */
     private long mExpire = DEFAULT_EXPIRE;
 
-    /** 日付のフォーマット. */
-    private SimpleDateFormat mSimpleDateFormat = new SimpleDateFormat("yyyyMMdd_kkmmss", Locale.JAPAN);
     /** Edit Image Thread. */
     private ExecutorService mImageService = Executors.newSingleThreadExecutor();
+
     private final DConnectApi mDrawImageApi = new PostApi() {
 
         @Override
@@ -55,7 +56,6 @@ public class HostCanvasProfile extends CanvasProfile {
 
         @Override
         public boolean onRequest(final Intent request, final Intent response) {
-
             String mode = getMode(request);
             String mimeType = getMIMEType(request);
             final CanvasDrawImageObject.Mode enumMode = CanvasDrawImageObject.convertMode(mode);
@@ -69,6 +69,7 @@ public class HostCanvasProfile extends CanvasProfile {
                     "Unsupported mimeType: " + mimeType);
                 return true;
             }
+
             final byte[] data = getData(request);
             final String uri = getURI(request);
             final double x = getX(request);
@@ -88,38 +89,13 @@ public class HostCanvasProfile extends CanvasProfile {
                 mImageService.execute(new Runnable() {
                     @Override
                     public void run() {
-                            sendImage(data, response, enumMode, x, y);
+                        sendImage(data, response, enumMode, x, y);
                     }
                 });
                 return false;
             }
         }
     };
-
-    /**
-     * Send Image.
-     * @param data binary
-     * @param response response message
-     * @param enumMode image mode
-     * @param x position
-     * @param y position
-     */
-    private void sendImage(byte[] data, Intent response, CanvasDrawImageObject.Mode enumMode, double x, double y) {
-        String uri = null;
-        try {
-            uri = writeForImage(data);
-        } catch (OutOfMemoryError e) {
-            MessageUtils.setIllegalDeviceStateError(response, e.getMessage());
-            sendResponse(response);
-            return;
-        } catch (IOException e) {
-            MessageUtils.setIllegalDeviceStateError(response, e.getMessage());
-            sendResponse(response);
-            return;
-        }
-        drawImage(response, uri, enumMode, x, y);
-        sendResponse(response);
-    }
 
     private final DConnectApi mDeleteImageApi = new DeleteApi() {
 
@@ -149,6 +125,25 @@ public class HostCanvasProfile extends CanvasProfile {
     public HostCanvasProfile() {
         addApi(mDrawImageApi);
         addApi(mDeleteImageApi);
+    }
+
+    /**
+     * Send Image.
+     * @param data binary
+     * @param response response message
+     * @param enumMode image mode
+     * @param x position
+     * @param y position
+     */
+    private void sendImage(byte[] data, Intent response, CanvasDrawImageObject.Mode enumMode, double x, double y) {
+        try {
+            drawImage(response, writeForImage(data), enumMode, x, y);
+        } catch (OutOfMemoryError e) {
+            MessageUtils.setIllegalDeviceStateError(response, e.getMessage());
+        } catch (IOException e) {
+            MessageUtils.setIllegalDeviceStateError(response, e.getMessage());
+        }
+        sendResponse(response);
     }
 
     /**
@@ -187,14 +182,6 @@ public class HostCanvasProfile extends CanvasProfile {
         ActivityManager activityMgr = (ActivityManager) getContext().getSystemService(Service.ACTIVITY_SERVICE);
         return activityMgr.getRunningTasks(1).get(0).topActivity.getClassName();
     }
-    /**
-     * 新規のファイル名を作成する.
-     *
-     * @return ファイル名
-     */
-    private String createNewFileName() {
-        return "android_canvas_" + mSimpleDateFormat.format(new Date());
-    }
 
     /**
      * 画像の保存
@@ -207,15 +194,11 @@ public class HostCanvasProfile extends CanvasProfile {
         File file = getContext().getCacheDir();
         FileOutputStream out = null;
         checkAndRemove(file);
-        File dstFile = File.createTempFile(createNewFileName(), "tmp", file);
+        File dstFile = File.createTempFile(CANVAS_PREFIX, ".tmp", file);
         try {
             out = new FileOutputStream(dstFile);
             out.write(data);
             out.close();
-        } catch (OutOfMemoryError e) {
-            throw new OutOfMemoryError(e.toString());
-        } catch (IOException e) {
-            e.printStackTrace();
         } finally {
             if (out != null) {
                 try {
@@ -238,39 +221,11 @@ public class HostCanvasProfile extends CanvasProfile {
             for (File childFile : file.listFiles()) {
                 checkAndRemove(childFile);
             }
-        } else if (file.isFile()) {
+        } else if (file.isFile() && file.getName().startsWith(CANVAS_PREFIX)) {
             long modified = file.lastModified();
             if (System.currentTimeMillis() - modified > mExpire) {
                 file.delete();
             }
-        }
-    }
-    /**
-     * ファイルをチェックして、中身を削除する.
-     *
-     * @param file 削除するファイル
-     */
-    private boolean checkAndRemoveInternal(@NonNull final File file) {
-        if (file.isDirectory()) {
-            for (File childFile : file.listFiles()) {
-                if (!checkAndRemoveInternal(childFile)) {
-                    return false;
-                }
-            }
-            return true;
-        } else if (file.isFile()) {
-            long modified = file.lastModified();
-            if (System.currentTimeMillis() - modified > mExpire) {
-                if (file.delete()) {
-                    return true;
-                } else {
-                    return false;
-                }
-            } else {
-                return true;
-            }
-        } else {
-            return false;
         }
     }
 }
